+++ source/Public/New-DscSelfSignedCertificate.ps1
@@ -55,8 +55,8 @@ function New-DscSelfSignedCertificate
                 There are build workers still on Windows Server 2012 R2 so let's
                 use the alternate method of New-SelfSignedCertificate.
             #>
-            Install-Module -Name PSPKI -Scope CurrentUser
-            Import-Module -Name PSPKI
+            # If you use this, declare PSPKI in RequiredModules, or install it
+            Import-Module -Name PSPKI -ErrorAction Stop
 
             $newSelfSignedCertificateExParameters = @{
                 Subject            = "CN=$certificateSubject"
@@ -75,11 +75,11 @@ function New-DscSelfSignedCertificate
             $certificate = New-SelfSignedCertificateEx @newSelfSignedCertificateExParameters
         }
 
-        Write-Info -Message ('Created self-signed certificate ''{0}'' with thumbprint ''{1}''.' -f $certificate.Subject, $certificate.Thumbprint)
+        Write-Verbose -Message ('Created self-signed certificate ''{0}'' with thumbprint ''{1}''.' -f $certificate.Subject, $certificate.Thumbprint)
     }
     else
     {
-        Write-Info -Message ('Using self-signed certificate ''{0}'' with thumbprint ''{1}''.' -f $certificate.Subject, $certificate.Thumbprint)
+        Write-Verbose -Message ('Using self-signed certificate ''{0}'' with thumbprint ''{1}''.' -f $certificate.Subject, $certificate.Thumbprint)
     }
 
     # Export the public key certificate
@@ -87,11 +87,11 @@ function New-DscSelfSignedCertificate
 
     # Update a machine and session environment variable with the path to the public certificate.
     Set-EnvironmentVariable -Name 'DscPublicCertificatePath' -Value $dscPublicCertificatePath -Machine
-    Write-Info -Message ('Environment variable $env:DscPublicCertificatePath set to ''{0}''' -f $env:DscPublicCertificatePath)
+    Write-Verbose -Message ('Environment variable $env:DscPublicCertificatePath set to ''{0}''' -f $env:DscPublicCertificatePath)
 
     # Update a machine and session environment variable with the thumbprint of the certificate.
     Set-EnvironmentVariable -Name 'DscCertificateThumbprint' -Value $certificate.Thumbprint -Machine
-    Write-Info -Message ('Environment variable $env:DscCertificateThumbprint set to ''{0}''' -f $env:DscCertificateThumbprint)
+    Write-Verbose -Message ('Environment variable $env:DscCertificateThumbprint set to ''{0}''' -f $env:DscCertificateThumbprint)
 
     return $certificate
 }