+++ tests/Unit/Private/Set-EnvironmentVariable.Tests.ps1
@@ -1,0 +1,17 @@
+$ProjectPath = "$PSScriptRoot\..\..\.." | Convert-Path
+$ProjectName = ((Get-ChildItem -Path $ProjectPath\*\*.psd1).Where{
+        ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) -and
+        $(try { Test-ModuleManifest $_.FullName -ErrorAction Stop } catch { $false } )
+    }).BaseName
+
+
+Import-Module $ProjectName -Force
+
+InModuleScope $ProjectName {
+    Describe 'Set-EnvironmentVariable' {
+        It 'Should be tested' {
+
+        }
+    }
+
+}